+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2023-12-10 - Feature: Allow the admin to hide the manual login form and the IDP login intro, solves #490.
 * 2023-12-10 - Improvement: Allow the admin to change the look of the course overview block, solves #204
 
 ### v4.3-r3